Changed the STL containers to QT containers, and replaced the size_t loops with int loops.  Still compiles without warnings.
git-svn-id: https://svn.code.sf.net/p/freemat/code/trunk@2928 58e251e6-400e-0410-b350-93cf1ec51ee9

+++ FreeMat2/libs/libCore/IO.cpp
@@ -1191,8 +1191,8 @@ ArrayVector FreadFunction(int nargout, const ArrayVector& arg) {
   // Next, we allocate space for the result
   void *qp = Malloc(elementCount*elementSize);
   // Read in the requested number of data points...
-  size_t g = fread(qp,elementSize,elementCount,fptr->fp);
-  if (g != (size_t)elementCount)
+  int g = fread(qp,elementSize,elementCount,fptr->fp);
+  if (g != elementCount)
     throw Exception("Insufficient data remaining in file to fill out requested size");
   if (ferror(fptr->fp)) 
     throw Exception(strerror(errno));
