+++ FreeMat2/libs/libCore/Random.cpp
@@ -132,7 +132,7 @@ ArrayVector RandBetaFunction(int nargout, const ArrayVector& arg) {
   float *dp = (float *) Malloc(sizeof(float)*outDims.getElementCount());
   float *p1 = (float*) arg1.getDataPointer();
   float *p2 = (float*) arg2.getDataPointer();
-  for (size_t i=0;i<outDims.getElementCount();i++) 
+  for (int i=0;i<outDims.getElementCount();i++) 
     dp[i] = genbet(p1[i*arg1_advance],p2[i*arg2_advance]);
   return singleArrayVector(Array(FM_FLOAT,outDims,dp));
 }
@@ -183,7 +183,7 @@ ArrayVector RandIFunction(int nargout, const ArrayVector& arg) {
   int32 *dp = (int32 *) Malloc(sizeof(int32)*outDims.getElementCount());
   int32 *p1 = (int32*) arg1.getDataPointer();
   int32 *p2 = (int32*) arg2.getDataPointer();
-  for (size_t i=0;i<outDims.getElementCount();i++) 
+  for (int i=0;i<outDims.getElementCount();i++) 
     dp[i] = ignuin(p1[i*arg1_advance],p2[i*arg2_advance]);
   return singleArrayVector(Array(FM_INT32,outDims,dp));
 }
@@ -238,7 +238,7 @@ ArrayVector RandChiFunction(int nargout, const ArrayVector& arg) {
   Dimensions outDims(arg1.dimensions());
   float *dp = (float *) Malloc(sizeof(float)*outDims.getElementCount());
   float *p1 = (float*) arg1.getDataPointer();
-  for (size_t i=0;i<outDims.getElementCount();i++) {
+  for (int i=0;i<outDims.getElementCount();i++) {
     if (p1[i] <= 0)
       throw Exception("argument to randchi must be positive");
     dp[i] = genchi(p1[i]);
@@ -282,7 +282,7 @@ ArrayVector RandExpFunction(int nargout, const ArrayVector& arg) {
   Dimensions outDims(arg1.dimensions());
   float *dp = (float *) Malloc(sizeof(float)*outDims.getElementCount());
   float *p1 = (float*) arg1.getDataPointer();
-  for (size_t i=0;i<outDims.getElementCount();i++) 
+  for (int i=0;i<outDims.getElementCount();i++) 
     dp[i] = genexp(p1[i]);
   return singleArrayVector(Array(FM_FLOAT,outDims,dp));
 }
@@ -327,7 +327,7 @@ ArrayVector RandPoissonFunction(int nargout, const ArrayVector& arg) {
   Dimensions outDims(arg1.dimensions());
   int32 *dp = (int32 *) Malloc(sizeof(int32)*outDims.getElementCount());
   float *p1 = (float*) arg1.getDataPointer();
-  for (size_t i=0;i<outDims.getElementCount();i++) 
+  for (int i=0;i<outDims.getElementCount();i++) 
     dp[i] = ignpoi(p1[i]);
   return singleArrayVector(Array(FM_INT32,outDims,dp));
 }
@@ -387,7 +387,7 @@ ArrayVector RandBinFunction(int nargout, const ArrayVector& arg) {
   uint32 *dp = (uint32 *) Malloc(sizeof(uint32)*outDims.getElementCount());
   uint32 *p1 = (uint32*) arg1.getDataPointer();
   float *p2 = (float*) arg2.getDataPointer();
-  for (size_t i=0;i<outDims.getElementCount();i++) 
+  for (int i=0;i<outDims.getElementCount();i++) 
     dp[i] = ignbin(p1[i*arg1_advance],p2[i*arg2_advance]);
   return singleArrayVector(Array(FM_UINT32,outDims,dp));
 }
@@ -510,7 +510,7 @@ ArrayVector RandFFunction(int nargout, const ArrayVector& arg) {
   float *dp = (float *) Malloc(sizeof(float)*outDims.getElementCount());
   float *p1 = (float*) arg1.getDataPointer();
   float *p2 = (float*) arg2.getDataPointer();
-  for (size_t i=0;i<outDims.getElementCount();i++) {
+  for (int i=0;i<outDims.getElementCount();i++) {
     if ((p1[i*arg1_advance] <= 0) || (p2[i*arg2_advance]) <= 0)
       throw Exception("randf requires positive arguments");
     dp[i] = genf(p1[i*arg1_advance],p2[i*arg2_advance]);
@@ -578,7 +578,7 @@ ArrayVector RandGammaFunction(int nargout, const ArrayVector& arg) {
   float *dp = (float *) Malloc(sizeof(float)*outDims.getElementCount());
   float *p1 = (float*) arg1.getDataPointer();
   float *p2 = (float*) arg2.getDataPointer();
-  for (size_t i=0;i<outDims.getElementCount();i++) {
+  for (int i=0;i<outDims.getElementCount();i++) {
     dp[i] = gengam(p1[i*arg1_advance],p2[i*arg2_advance]);
   }
   return singleArrayVector(Array(FM_FLOAT,outDims,dp));
@@ -634,12 +634,12 @@ ArrayVector RandMultiFunction(int nargout, const ArrayVector& arg) {
   float *dp;
   dp = (float*) arg2.getReadWriteDataPointer();
   float Psum = 0.0;
-  for (size_t i=0;i<arg2.getLength();i++) {
+  for (int i=0;i<arg2.getLength();i++) {
     if ((dp[i] < 0) || (dp[i] > 1)) 
       throw Exception("probabiliy vector argument to randmulti must have all elements between 0 and 1");
     Psum += dp[i];
   }
-  for (size_t i=0;i<arg2.getLength();i++) {
+  for (int i=0;i<arg2.getLength();i++) {
     dp[i] /= Psum;
   }
   Dimensions outDims;
@@ -698,7 +698,7 @@ ArrayVector RandNChiFunction(int nargout, const ArrayVector& arg) {
   float *dp = (float *) Malloc(sizeof(float)*outDims.getElementCount());
   float *p1 = (float*) arg1.getDataPointer();
   float *p2 = (float*) arg2.getDataPointer();
-  for (size_t i=0;i<outDims.getElementCount();i++) {
+  for (int i=0;i<outDims.getElementCount();i++) {
     if (p1[i*arg1_advance] <= 1.0)
       throw Exception("degrees of freedom argument must be > 1.0");
     if (p2[i*arg2_advance] < 0.0)
@@ -770,7 +770,7 @@ ArrayVector RandNFFunction(int nargout, const ArrayVector& arg) {
   float *p1 = (float*) arg1.getDataPointer();
   float *p2 = (float*) arg2.getDataPointer();
   float *p3 = (float*) arg3.getDataPointer();
-  for (size_t i=0;i<outDims.getElementCount();i++) {
+  for (int i=0;i<outDims.getElementCount();i++) {
     if (p1[i*arg1_advance] <= 1.0)
       throw Exception("numerator degrees of freedom argument must be > 1.0");
     if (p2[i*arg2_advance] <= 0.0)
